+++ src/com/kaylerrenslow/armaDialogCreator/arma/control/impl/StaticControl.java
@@ -82,7 +82,8 @@ public class StaticControl extends ArmaControl {
 	};
 
 	public StaticControl(@NotNull String name, int idc, @NotNull ArmaResolution resolution, @NotNull Env env, @NotNull SpecificationRegistry registry) {
-		super(ControlType.Static, name, SPEC_PROVIDER, idc, ControlStyle.CENTER.getStyleGroup(), resolution, RendererLookup.Static, env, registry);
+		super(ControlType.Static, name, SPEC_PROVIDER, idc, resolution, RendererLookup.Static, env, registry);
+		findProperty(ControlPropertyLookup.STYLE).setValueIfAbsent(true, ControlStyle.CENTER.getStyleGroup());
 	}
 
 }