+++ src/com/kaylerrenslow/armaDialogCreator/gui/main/controlPropertiesEditor/ControlStyleValueEditor.java
@@ -43,13 +43,13 @@ public class ControlStyleValueEditor extends HBox implements ValueEditor<Control
 		menuButton.getSelectedItems().addListener(new ListChangeListener<ControlStyle>() {
 			@Override
 			public void onChanged(Change<? extends ControlStyle> c) {
-				String s = "";
+				StringBuilder s = new StringBuilder();
 				List<ControlStyle> selected = menuButton.getSelectedItems();
 				for (int i = 0; i < selected.size(); i++) {
-					s += selected.get(i).styleValue + (i != selected.size() - 1 ? ControlStyleGroup.DEFAULT_DELIMITER : "");
+					s.append(selected.get(i).styleValue).append(i != selected.size() - 1 ? ControlStyleGroup.DEFAULT_DELIMITER : "");
 				}
-				textField.setText(s);
-				ControlStyleGroup group = menuButton.getSelectedItems().size() == 0 ? null :
+				textField.setText(s.toString());
+				ControlStyleGroup group = menuButton.getSelectedItems().size() == 0 ? new ControlStyleGroup(ControlStyle.EMPTY) :
 						new ControlStyleGroup(menuButton.getSelectedItems().toArray(new ControlStyle[menuButton.getSelectedItems().size()]));
 				valueObserver.updateValue(group);
 			}